+++ src/universal/components/NullCard/NullCard.js
@@ -10,10 +10,11 @@ import makeUsername from 'universal/utils/makeUsername';
 const NullCard = (props) => {
   const {styles, type, preferredName} = props;
   const username = makeUsername(preferredName);
+  const label = type === 'Action' ? 'an Action' : 'a Project';
   return (
     <div className={css(styles.root)}>
       <Type align="center" bold scale="s3" colorPalette="mid">
-        @{username}<br />is adding a {type}<Ellipsis />
+        @{username}<br />is adding {label}<Ellipsis />
       </Type>
     </div>
   );