+++ src/universal/modules/meeting/components/CheckinCard/CheckinCard.js
@@ -3,11 +3,11 @@ import withStyles from 'universal/styles/withStyles';
 import {css} from 'aphrodite-local-styles/no-important';
 import appTheme from 'universal/styles/theme/appTheme';
 import Avatar from 'universal/components/Avatar/Avatar';
-import checkInCardBaseStyles from './checkInCardBaseStyles';
-import CheckInCardButtons from 'universal/modules/meeting/components/CheckInCardButtons/CheckInCardButtons';
+import CheckinCardBaseStyles from './CheckinCardBaseStyles';
+import CheckinCardButtons from 'universal/modules/meeting/components/CheckinCardButtons/CheckinCardButtons';
 import {withRouter} from 'react-router';
 
-const CheckInCard = (props) => {
+const CheckinCard = (props) => {
   const {
     checkInPressFactory,
     handleCardClick,
@@ -27,12 +27,12 @@ const CheckInCard = (props) => {
     <div className={cardStyles} onClick={!isActive && handleCardClick}>
       <Avatar {...member} size="largest"/>
       <div className={css(styles.cardName)}>{preferredName}</div>
-      {isActive && <CheckInCardButtons checkInPressFactory={checkInPressFactory} isCheckedIn={isCheckedIn}/>}
+      {isActive && <CheckinCardButtons checkInPressFactory={checkInPressFactory} isCheckedIn={isCheckedIn}/>}
     </div>
   );
 };
 
-CheckInCard.propTypes = {
+CheckinCard.propTypes = {
   checkInPressFactory: PropTypes.func,
   handleCardClick: PropTypes.func,
   isActive: PropTypes.bool,
@@ -42,7 +42,7 @@ CheckInCard.propTypes = {
 
 const styleThunk = () => ({
   card: {
-    ...checkInCardBaseStyles,
+    ...CheckinCardBaseStyles,
     borderColor: appTheme.palette.mid50l
   },
 
@@ -65,4 +65,4 @@ const styleThunk = () => ({
   }
 });
 
-export default withRouter(withStyles(styleThunk)(CheckInCard));
+export default withRouter(withStyles(styleThunk)(CheckinCard));